move GraphAssertions to common test lib
Avoid code duplication for GraphAssertions helper. To
avoid a cyclic dependency between chart and test libs
the PngImage utility was moved to core. That utility
has no dependencies other than the jdk.
