+++ atlas-chart/src/main/scala/com/netflix/atlas/chart/PngGraphEngine.scala
@@ -18,6 +18,8 @@ package com.netflix.atlas.chart
 import java.awt.image.RenderedImage
 import java.io.OutputStream
 
+import com.netflix.atlas.core.util.PngImage
+
 
 trait PngGraphEngine extends GraphEngine {
   val contentType: String = "image/png"
