+++ atlas-core/src/main/scala/com/netflix/atlas/core/util/PngImage.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.atlas.chart
+package com.netflix.atlas.core.util
 
 import java.awt.Color
 import java.awt.Font
