+++ atlas-core/src/test/resources/pngimage/test.png
