+++ atlas-core/src/test/resources/pngimage/test_diff.png
