+++ atlas-core/src/test/resources/pngimage/test_error_400x300.png
