+++ atlas-core/src/test/scala/com/netflix/atlas/core/util/PngImageSuite.scala
@@ -13,11 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.atlas.chart
+package com.netflix.atlas.core.util
 
 import java.io.InputStream
-
-import com.netflix.atlas.core.util.Streams
 import org.scalatest.FunSuite
